+++ discon/file_content.py
@@ -273,3 +273,26 @@ script:
 after_success:
     - coveralls
 """
+
+_TESTS_MAIN_PY = """\
+#! /usr/bin/python
+# -*- coding: utf-8 -*-
+
+# import logging
+# logger = logging.getLogger(__name__)
+from loguru import logger
+import pytest
+import os.path
+
+path_to_script = os.path.dirname(os.path.abspath(__file__))
+
+def inc(x):
+    return x + 1
+
+
+
+# @pytest.mark.interactive
+# @pytest.mark.slow
+def test_answer():
+    assert inc(3) == 5
+"""